+++ client/rpc/status.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/spf13/cobra"
 
+	"github.com/cosmos/cosmos-sdk/client/grpc/tmservice"
 	"github.com/tendermint/tendermint/libs/bytes"
 	"github.com/tendermint/tendermint/p2p"
 	coretypes "github.com/tendermint/tendermint/rpc/core/types"
@@ -26,9 +27,21 @@ type validatorInfo struct {
 // ResultStatus is node's info, same as Tendermint, except that we use our own
 // PubKey.
 type resultStatus struct {
-	NodeInfo      p2p.DefaultNodeInfo
+	NodeInfo      nodeInfo
 	SyncInfo      coretypes.SyncInfo
 	ValidatorInfo validatorInfo
+}
+
+type nodeInfo struct {
+	ProtocolVersion p2p.ProtocolVersion      `json:"protocol_version"`
+	DefaultNodeID   p2p.ID                   `json:"default_node_id,omitempty"`
+	ListenAddr      string                   `json:"listen_addr,omitempty"`
+	Network         string                   `json:"network,omitempty"`
+	Version         string                   `json:"version,omitempty"`
+	Channels        bytes.HexBytes           `json:"channels,omitempty"`
+	Moniker         string                   `json:"moniker,omitempty"`
+	Other           p2p.DefaultNodeInfoOther `json:"other"`
+	BinaryVersion   string                   `json:"binary_version"`
 }
 
 // StatusCommand returns the command to return the status of the network.
@@ -38,6 +51,12 @@ func StatusCommand() *cobra.Command {
 		Short: "Query remote node for status",
 		RunE: func(cmd *cobra.Command, _ []string) error {
 			clientCtx, err := client.GetClientQueryContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			queryClient := tmservice.NewServiceClient(clientCtx)
+			res, err := queryClient.GetNodeInfo(context.Background(), &tmservice.GetNodeInfoRequest{})
 			if err != nil {
 				return err
 			}
@@ -52,8 +71,21 @@ func StatusCommand() *cobra.Command {
 			if err != nil {
 				return err
 			}
+
+			infoWithBinary := nodeInfo{
+				ProtocolVersion: status.NodeInfo.ProtocolVersion,
+				DefaultNodeID:   status.NodeInfo.DefaultNodeID,
+				ListenAddr:      status.NodeInfo.ListenAddr,
+				Network:         status.NodeInfo.Network,
+				Version:         status.NodeInfo.Version,
+				Channels:        status.NodeInfo.Channels,
+				Moniker:         status.NodeInfo.Moniker,
+				Other:           status.NodeInfo.Other,
+				BinaryVersion:   res.ApplicationVersion.GetVersion(),
+			}
+
 			statusWithPk := resultStatus{
-				NodeInfo: status.NodeInfo,
+				NodeInfo: infoWithBinary,
 				SyncInfo: status.SyncInfo,
 				ValidatorInfo: validatorInfo{
 					Address:     status.ValidatorInfo.Address,